1_MaterialComplementario/gestor-ia/apps/api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.7.4
python-dotenv==1.0.1
redis==5.0.7
rq==1.16.2
pyjwt==2.8.0
cryptography==42.0.8
pgvector==0.3.2
python-multipart==0.0.9
